feat: refactor robots content
